compliant: ``new`` has new syntax
``new`` now takes only one argument, the constructor that must be called. The arguments of the constructor are passed in another call. For instance::

   new(Date)(1, 2, 3)

This looks better than the old syntax. Also the constructor arguments in this case ``1, 2, 3`` are run through Veloce, it means that it will call the constructor with JavaScript types. If it's not what you meant you can use the following trick::

  a = 1
  b = 2
  c = 3
  new(Date)(a, b, c)

In the above example a, b and c are Python types.

Also since Veloce know nothing about Pythonium compliant OO system. If you want to pass attributes to the constructor you need to do it outside of the call statment::

  a = some_object.method()
  b = some_other_object.property
  c = jscode('3')
  new(Date)(a, b, c)